Move FlowNode parser functions into separate modules
